+++ source/extensions/isaacsim.ros2.bridge/python/tests/test_ros2_odometry.py
@@ -31,7 +31,6 @@ from isaacsim.core.utils.physics import simulate_async
 from isaacsim.core.utils.prims import is_prim_path_valid
 from isaacsim.core.utils.stage import open_stage_async
 from isaacsim.core.utils.string import find_unique_string_name
-from isaacsim.storage.native import get_assets_root_path_async
 
 from .common import ROS2TestCase, get_qos_profile
 
@@ -42,11 +41,6 @@ class TestRos2Odometry(ROS2TestCase):
         await super().setUp()
 
         await omni.usd.get_context().new_stage_async()
-
-        self._assets_root_path = await get_assets_root_path_async()
-        if self._assets_root_path is None:
-            carb.log_error("Could not find Isaac Sim assets folder")
-            return
 
         self.CUBE_SCALE = 0.5
         await omni.kit.app.get_app().next_update_async()