+++ source/extensions/isaacsim.test.utils/isaacsim/test/utils/__init__.py
@@ -13,4 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .image_capture import *
 from .image_comparison import *
+from .timed_async_test import *