+++ source/extensions/isaacsim.ros2.bridge/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.12.1] - 2025-09-08
+### Changed
+- Use base test class from isaacsim.test.utils 
+- Reduce test time by using fast asset root path function
+
 ## [4.12.0] - 2025-09-05
 ### Changed
 - Update internal ROS 2 Humble and Jazzy libs to include common_interfaces, tf2_ros, sensor_msgs_py, Simulation Interfaces v1.1.0.